generic alphabetize
